requirements.txt
streamlit
pandas
numpy
openpyxl
pdfplumber
google-generativeai
xlrd
